Moved device tracker functions into device tracker
